the alarm can be triggered now

+++ app/src/main/java/com/example/crossclock/data/alarm/Alarm.kt
@@ -9,9 +9,8 @@ import java.time.ZoneId
 @Entity
 data class Alarm(
     @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
+    val id: Int = 0,
     val time: LocalDateTime,
     val message: String,
     val timeZone: ZoneId,
-    //val date: String
 )
